Backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.5.0
PyJWT==2.8.0